+++ src/math/add_custom/op_host/add_custom_tiling.h
@@ -17,14 +17,14 @@
 
 namespace optiling {
 BEGIN_TILING_DATA_DEF(TilingData)
-  TILING_DATA_FIELD_DEF(uint32_t, smallCoreDataNum);
-  TILING_DATA_FIELD_DEF(uint32_t, bigCoreDataNum);
-  TILING_DATA_FIELD_DEF(uint32_t, ubPartDataNum);
-  TILING_DATA_FIELD_DEF(uint32_t, smallCoreTailDataNum);
-  TILING_DATA_FIELD_DEF(uint32_t, bigCoreTailDataNum);
-  TILING_DATA_FIELD_DEF(uint32_t, smallCoreLoopNum);
-  TILING_DATA_FIELD_DEF(uint32_t, bigCoreLoopNum);
-  TILING_DATA_FIELD_DEF(uint32_t, tailBlockNum);
+  TILING_DATA_FIELD_DEF(uint64_t, smallCoreDataNum);
+  TILING_DATA_FIELD_DEF(uint64_t, bigCoreDataNum);
+  TILING_DATA_FIELD_DEF(uint64_t, ubPartDataNum);
+  TILING_DATA_FIELD_DEF(uint64_t, smallCoreTailDataNum);
+  TILING_DATA_FIELD_DEF(uint64_t, bigCoreTailDataNum);
+  TILING_DATA_FIELD_DEF(uint64_t, smallCoreLoopNum);
+  TILING_DATA_FIELD_DEF(uint64_t, bigCoreLoopNum);
+  TILING_DATA_FIELD_DEF(uint64_t, tailBlockNum);
 END_TILING_DATA_DEF;
 
 REGISTER_TILING_DATA_CLASS(AddCustom, TilingData)
